requirements.txt
# Contact Center Agent System - Python Dependencies

# Core AI and Agent Framework
openai>=1.12.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
langgraph-supervisor

# MongoDB
pymongo>=4.6.0
motor>=3.3.0

# MCP (Model Context Protocol)
mcp>=0.1.0

# Data Processing
pydantic>=2.5.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.9.0
